Rewordings; improve venv usage in devops scripts

+++ devops/db_recreate.sh
@@ -3,9 +3,9 @@
 set -e
 
 pushd "$(dirname $0)/.."
-source venv/bin/activate
-rm -i db.sqlite3 || true
+[[ -f db.sqlite3 ]] && rm -i db.sqlite3 || true
+venv/bin/python ./manage.py flush --no-input
 devops/db_create.sh
 echo "Creating superuser..."
-python3 ./manage.py createsuperuser --username admin --email noreply@joanardiaca.net
+venv/bin/python ./manage.py createsuperuser --username admin --email noreply@joanardiaca.net
 popd
